TASK-3962 Renaming: playerId to playerName, don't do player lookup on creation
It's good that JuicyRaspberryPie takes the names, as the UUID lookup would make API use a bit more complex (especially for absolute beginners). The mapping is practically stable, anyway.
As we don't need the UUID, there's also little sense to do the lookup immediately after connecting. This would just throw an error, but the underlying player might not actually be used.
As players can enter and leave throughout the program's runtime, it makes sense to handle this lazily - it doesn't make sense to enforce the existence of the passed player eagerly when they could have vanished once it is actually used.