TASK-3962 FIX: Minecraft.setPlayer() also persists the player locally, getPlayerName() returns that
JuicyRaspberryPie has a notion of attachedPlayer that is set by setPlayer() and defaults to the first active user when the connection is established.
Let's implement this logic: If no player is passed to mcpi, use that first active user; if that user leaves, switch to another available user (or fail when noone's there). In this case, Minecraft.getPlayerName() always needs to query the server for the current one, and its Minecraft.playerName remains empty.
If a player is passed or later set via Minecraft.setPlayer(), the instance will be linked to that; the player is stored in Minecraft.playerName; Minecraft.getPlayerName() doesn't need to ask the server. If that player leaves, any player-related calls will fail until the player returns.