TASK-3962 Make Minecraft.getPlayerName() on floating instance return the current player name
If no playerName argument is passed to Minecraft.create(), the default [] is stored in Minecraft.playerName, which is unexpected. The user shouldn't have to invoke Minecraft.getPlayerName() to obtain the actual player.
Refactor Minecraft.playerName into a read-only property that delegates to Minecraft.getPlayerName(), and store the actual value in the _playerName attribute now.