TASK-4017 FIX: Minecraft.getPlayerEntityIds() should just return the UUIDs
Given that we already have Minecraft.getPlayerNames(), there might be a need for a method that returns the UUID → playerName mapping. But getPlayerEntityIds() should not return unparsed UUID:playerName tuples. The entity ID by itself is useful in combination with the Entity class.
The method also is documented to just return the [id].