TASK-4017 Refactoring: Extract Connection.sendReceiveObjectList() and replace the individual listcomps with it
The CmdEvents.poll*() methods and Minecraft.get{NearbyEntities,Player{EntityIds,Names}}() all do splitting on a single object's marshaled representation and then pick or pass the resulting attributes.
Now that the separation (| for objects, comma for an object's attributes) has been aligned, a generic method can do that in each case.
We just a tiny specialization for CmdEvents.pollChatPosts() in that the splitting must stop after 2; by accepting kwargs and forwarding those to str.split(), the maxsplit can be customized.